requirements.txt
requests
ruamel.yaml
pydantic